requirements.txt
torch~=1.5.0
numpy~=1.18.1
Pillow~=7.1.2
torchvision~=0.6.0
mlflow~=1.7.2
tqdm~=4.43.0
dominate~=2.4.0
pandas~=1.0.0
matplotlib~=3.1.3
opencv-python~=4.2.0.32
scipy~=1.4.1
kornia~=0.2.0
imageio~=2.6.1
networkx~=2.4
scikit-image~=0.16.2
sklearn~=0.0
scikit-learn~=0.22.1